+++ src/4_Shared/lib/dateFormatter.ts
@@ -28,3 +28,25 @@ export const formatTime30 = Array.from({ length: 48 }, (_, i) => {
   const minute = i % 2 === 0 ? "00" : "30";
   return `${hour}:${minute}`;
 });
+
+//  * 날짜를 표시용 문자열로 포맷합니다 X월 X일
+export const formatDateForDisplay = (date: Date): string => {
+  const today = new Date();
+  const inputNormalized = new Date(date);
+
+  const tomorrow = new Date(today);
+  tomorrow.setDate(today.getDate() + 1);
+
+  const yesterday = new Date(today);
+  yesterday.setDate(today.getDate() - 1);
+
+  if (inputNormalized.getTime() === today.getTime()) {
+    return "오늘";
+  } else if (inputNormalized.getTime() === tomorrow.getTime()) {
+    return "내일";
+  } else if (inputNormalized.getTime() === yesterday.getTime()) {
+    return "어제";
+  }
+
+  return `${inputNormalized.getMonth() + 1}월 ${inputNormalized.getDate()}일`;
+};
