+++ src/4_Shared/lib/dateUtil.ts
@@ -1,0 +1,9 @@
+//  * 특정 날짜에 매치가 있는지 확인합니다
+
+export const isSameDate = (date1: Date, date2: Date): boolean => {
+  return (
+    date1.getFullYear() === date2.getFullYear() &&
+    date1.getMonth() === date2.getMonth() &&
+    date1.getDate() === date2.getDate()
+  );
+};
